Add basic floating windows

+++ WindowSwitcherLib/Models/WindowConfig.cs
@@ -5,8 +5,8 @@ public class WindowConfig
     public string WindowId { get; set; }
     public string WindowTitle { get; set; }
     public string ShortWindowTitle { get; set; }
-    public int WindowWidth { get; set; } = 60;
-    public int WindowHeight { get; set; } = 40;
+    public double WindowWidth { get; set; } = 60;
+    public double WindowHeight { get; set; } = 40;
     public int WindowLeft { get; set; } = 100;
     public int WindowTop { get; set; } = 100;
 }
